Start implementing (de)serialisation.
Having lexers defined only in Go is not ideal for a couple of reasons.
Firstly, it impedes a significant portion of contributors who use Chroma
in Hugo, but don't know Go. Secondly, it bloats the binary size of any
project that imports Chroma.

Here's an in-progress example for the pkgconfig lexer:

    <rules>
      <state name="spvalue">
        <rule>
          <include state="interp"/>
        </rule>
        <rule pattern="#.*$">
          <token type="CommentSingle"/>
          <pop depth="1"/>
        </rule>
        <rule pattern="\n">
          <token type="Text"/>
          <pop depth="1"/>
        </rule>
        <rule pattern="[^${}#\n]+">
          <token type="Text"/>
        </rule>
        <rule pattern=".">
          <token type="Text"/>
        </rule>
      </state>
      <state name="root">
        <rule pattern="#.*$">
          <token type="CommentSingle"/>
        </rule>
        <rule pattern="^(\w+)(=)">
          <bygroups emit="2001" emit="4000"/>
        </rule>
        <rule pattern="^([\w.]+)(:)">
          <bygroups emit="2018" emit="5000"/>
          <push state="spvalue"/>
        </rule>
        <rule>
          <include state="interp"/>
        </rule>
        <rule pattern="[^${}#=:\n.]+">
          <token type="Text"/>
        </rule>
        <rule pattern=".">
          <token type="Text"/>
        </rule>
      </state>
      <state name="interp">
        <rule pattern="\$\$">
          <token type="Text"/>
        </rule>
        <rule pattern="\$\{">
          <token type="LiteralStringInterpol"/>
          <push state="curly"/>
        </rule>
      </state>
      <state name="curly">
        <rule pattern="\}">
          <token type="LiteralStringInterpol"/>
          <pop depth="1"/>
        </rule>
        <rule pattern="\w+">
          <token type="NameAttribute"/>
        </rule>
      </state>
    </rules>

Why XML? YAML is an abomination and JSON is not human editable. XML
also compresses very well (eg. Go template lexer XML compresses from
3239 bytes to 718).

Why a new syntax format? All major existing formats rely on the
Oniguruma regex engine, which is extremely complex and for which there
is no Go port.

Why not earlier? Prior to the existence of fs.FS this was not a viable
option.